+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement `ElementFixedLength` for `Option<T>` where `T: ElementFixedLength`
 - Add chunk compaction API to remove extraneous bytes from encoded chunks
   - Add `Array::[async_]compact_chunk()`
-  - Add `{Array,Bytes}ToBytesCodecTraits::compact()` with a default implementation
+  - Add `ArrayToBytesCodecTraits::compact()` with a default implementation
 
 ### Changed